+++ src/javascript/display/menu.js
@@ -26,6 +26,13 @@ Menu.prototype.addControl = function(){
     let element = document.createElement( "div" );
     element.classList.add( "gadget-ui-menu-item" );
     element.innerText = label;
+    let image = ( item.image !== undefined ? item.image : "" );
+    if( image.length ){
+      let imgEl = document.createElement( "img" );
+      imgEl.src = image;
+      imgEl.classList.add( "gadget-ui-menu-icon" );
+      element.appendChild( imgEl );
+    }
     if( item.link !== undefined && item.link !== null && ( item.link.length > 0 || typeof item.link === 'function' ) ){
       //element.removeEventListener( "click" );
       element.style.cursor = 'pointer';
@@ -59,7 +66,15 @@ Menu.prototype.addControl = function(){
     //let element = `<div class="gadget-ui-menu">{menuData.label}</div>`;
     let element = document.createElement( "div" );
     element.classList.add( "gadget-ui-menu" );
-    element.innerText = menuData.label;
+    let label = ( menuData.label !== undefined ? menuData.label : "" );
+    element.innerText = label;
+    let image = ( menuData.image !== undefined ? menuData.image : "" );
+    if( image.length ){
+      let imgEl = document.createElement( "img" );
+      imgEl.src = image;
+      imgEl.classList.add( "gadget-ui-menu-icon" );
+      element.appendChild( imgEl );
+    }
     // process the menuItem
     element.appendChild( processMenuItem( menuData.menuItem, element ) );
     return element;